requirements.txt
fastapi[standard]
pydantic
motor
passlib[bcrypt]
python-jose
python-dotenv